More clean ups
